backend/requirements.txt
gradio==4.44.0
fastapi==0.104.1
torch==2.1.0
torchvision==0.16.0
tensorflow>=2.15.0
pillow==10.0.1
numpy>=1.24.3
python-multipart>=0.0.9 